refactor and break apart
