+++ client/app/admin/inventory/_components/RemoveItemForm.tsx
@@ -1,0 +1,5 @@
+"use client";
+
+export function RemoveItemForm() {
+  return <div>remove item</div>;
+}
